Refactor tournament admin UI into setup, ops, and live pages
