+++ jupr_app/ui/page_registry.py
@@ -38,7 +38,9 @@ PAGE_DEFINITIONS: tuple[PageDefinition, ...] = (
     PageDefinition("jupr_live_admin", "🔴 JUPR Live Admin", admin_only=True),
     PageDefinition("theme_qa", "🎨 Theme QA", admin_only=True),
     PageDefinition("tournaments", "🏆 Tournaments", admin_only=True),
-    PageDefinition("tournament_manager", "🏆 Tournament Manager", admin_only=True),
+    PageDefinition("tournament_manager", "🛠️ Tournament Setup", admin_only=True),
+    PageDefinition("tournament_ops", "📋 Tournament Operations", admin_only=True),
+    PageDefinition("tournament_live", "🔴 Tournament Live", admin_only=True),
     PageDefinition("tournament_registration", "📝 Tournament Registration", public=True),
     PageDefinition("tournament_partner_board", "🤝 Partner Board", public=True),
     PageDefinition("weekly_recap", "🗞️ Weekly Recap", public=True),
