+++ jupr_app/ui/pages/tournament_ops.py
@@ -1,0 +1,1374 @@
+from __future__ import annotations
+
+from datetime import date, datetime
+from io import StringIO
+from typing import Any
+
+import pandas as pd
+import streamlit as st
+
+from jupr_app.domain.event_tags import derive_default_date_tags, normalize_event_tags
+from jupr_app.domain.match_processing import process_matches
+from jupr_app.domain.player_ops import safe_add_player
+from jupr_app.domain.tournament_match_payload import build_tournament_match_payload
+from jupr_app.domain.tournament_results_import import (
+    STAGE_OPTIONS,
+    build_draw_import_payload,
+    parse_dupr_results_csv,
+    suggest_player_matches,
+)
+from jupr_app.domain.tournaments import (
+    SUPPORTED_TEAM_COUNTS,
+    build_playoff_games,
+    build_podium_payload,
+    build_round_robin_games,
+    compute_podium_from_playoffs,
+    compute_podium_from_rr,
+    compute_round_robin_standings,
+    finalize_game,
+    resolve_playoff_dependencies,
+)
+from jupr_app.domain.tournament_podium import award_tournament_trophies_from_podium, upsert_tournament_podium
+from jupr_app.domain.tournament_registration_repo import (
+    archive_tournament,
+    build_public_urls,
+    build_registration_state,
+    delete_unused_draft_tournament,
+    get_registration_settings,
+    list_event_options as list_registration_event_options,
+    list_existing_tournaments,
+    list_registration_days,
+    registration_feature_available,
+    tournament_can_be_deleted,
+    unarchive_tournament,
+    upsert_registration_settings,
+    list_registrations,
+    update_registration_admin_fields,
+    ADMIN_PAYMENT_STATUS_OPTIONS,
+    ADMIN_REGISTRATION_STATUS_OPTIONS,
+)
+from jupr_app.ui.layout import page_shell
+
+LEGACY_DEFAULT_TEAM_COUNT = 4
+TOURNAMENT_STATUS_OPTIONS = ["DRAFT", "REGISTRATION", "REGISTRATION_OPEN", "REGISTRATION_CLOSED", "ARCHIVED"]
+TOURNAMENT_LOCALE_OPTIONS = ["en", "es", "bilingual"]
+IMPORT_SOURCE_OPTIONS = ["REGISTRATION", "MANUAL", "BULK_UPLOAD"]
+
+
+def _safe_text(value: Any) -> str:
+    return str(value or "").strip()
+
+
+def _normalize_name(value: object) -> str:
+    return " ".join(str(value or "").strip().lower().split())
+
+
+def _coerce_date(value: Any, fallback: date) -> date:
+    if isinstance(value, datetime):
+        return value.date()
+    if isinstance(value, date):
+        return value
+    return fallback
+
+
+def _parse_optional_date(value: Any) -> str | None:
+    if isinstance(value, date):
+        return value.isoformat()
+    text = _safe_text(value)
+    return text or None
+
+
+def _go_to_tournament_manager(tournament_id: Any) -> None:
+    st.query_params["page"] = "tournament_manager"
+    st.query_params["tournament_id"] = tournament_id
+    st.rerun()
+
+
+
+
+def _tournament_date_window_text(tournament: dict[str, Any]) -> str | None:
+    start = _safe_text(tournament.get("start_date"))
+    end = _safe_text(tournament.get("end_date"))
+    if not start or not end:
+        return None
+    return f"{start} → {end}"
+def _resolve_player_id(name: str, name_to_id: dict[str, Any], id_to_name: dict[Any, str]) -> Any:
+    if name in name_to_id:
+        return name_to_id[name]
+    normalized = _normalize_name(name)
+    if not normalized:
+        return None
+    for known_name, player_id in name_to_id.items():
+        if _normalize_name(known_name) == normalized:
+            return player_id
+    for player_id, known_name in id_to_name.items():
+        if _normalize_name(known_name) == normalized:
+            return player_id
+    return None
+
+
+def _find_event_roster(registration_bridge: dict | None, *, event_option_id: str, registration_day_id: str) -> dict | None:
+    state = (registration_bridge or {}).get("state") or {}
+    for roster in state.get("event_rosters") or []:
+        if str(roster.get("event_option_id")) == str(event_option_id) and str(roster.get("event_day_id")) == str(registration_day_id):
+            return roster
+    return None
+
+
+def _parse_bulk_upload(file, pasted_text: str) -> pd.DataFrame:
+    if file is not None:
+        name = str(getattr(file, "name", "")).lower()
+        if name.endswith(".xlsx"):
+            return pd.read_excel(file)
+        return pd.read_csv(file)
+    text = _safe_text(pasted_text)
+    if not text:
+        return pd.DataFrame()
+    sep = "\t" if "\t" in text and text.count("\t") >= text.count(",") else ","
+    return pd.read_csv(StringIO(text), sep=sep)
+
+
+def _canonical_import_df(raw_df: pd.DataFrame) -> pd.DataFrame:
+    if raw_df.empty:
+        return pd.DataFrame(columns=["Player 1", "Player 2", "Team Name", "Seed", "Notes"])
+    aliases = {
+        "player 1": "Player 1",
+        "player1": "Player 1",
+        "player one": "Player 1",
+        "player 2": "Player 2",
+        "player2": "Player 2",
+        "player two": "Player 2",
+        "team": "Team Name",
+        "team name": "Team Name",
+        "seed": "Seed",
+        "notes": "Notes",
+    }
+    renamed = {}
+    for col in raw_df.columns:
+        key = " ".join(str(col).strip().lower().split())
+        renamed[col] = aliases.get(key, str(col).strip())
+    df = raw_df.rename(columns=renamed).copy()
+    for col in ["Player 1", "Player 2", "Team Name", "Seed", "Notes"]:
+        if col not in df.columns:
+            df[col] = None
+    return df[["Player 1", "Player 2", "Team Name", "Seed", "Notes"]]
+
+
+def _insert_tournament_shell(supabase, payload: dict) -> None:
+    try:
+        supabase.table("tournaments").insert(payload).execute()
+        return
+    except Exception:
+        pass
+
+    fallback_payload = {
+        "club_id": payload["club_id"],
+        "name": payload["name"],
+        "status": payload.get("status") or "DRAFT",
+        "team_count": int(payload.get("team_count") or LEGACY_DEFAULT_TEAM_COUNT),
+    }
+    supabase.table("tournaments").insert(fallback_payload).execute()
+
+
+def _list_draws(supabase, tournament_id: str) -> list[dict[str, Any]]:
+    try:
+        resp = (
+            supabase.table("tournament_event_draws")
+            .select("*")
+            .eq("tournament_id", tournament_id)
+            .order("created_at", desc=False)
+            .execute()
+        )
+        return resp.data or []
+    except Exception:
+        return []
+
+
+def _create_draw(supabase, payload: dict) -> None:
+    supabase.table("tournament_event_draws").insert(payload).execute()
+
+
+def _update_ops_field(supabase, tournament_id: str, draw_id: str | None, field: str, value: Any) -> None:
+    if draw_id:
+        try:
+            supabase.table("tournament_event_draws").update({field: value}).eq("id", draw_id).execute()
+            return
+        except Exception:
+            pass
+    supabase.table("tournaments").update({field: value}).eq("id", tournament_id).execute()
+
+
+def _get_draw_setting(draw: dict[str, Any] | None, tournament: dict[str, Any], field: str, default: Any) -> Any:
+    if draw and draw.get(field) not in (None, "", "None"):
+        return draw.get(field)
+    if tournament.get(field) not in (None, "", "None"):
+        return tournament.get(field)
+    return default
+
+
+def _effective_team_count(draw: dict[str, Any] | None, tournament: dict[str, Any], teams_by_number: dict[int, dict]) -> int:
+    seeded = _get_draw_setting(draw, tournament, "team_count", None)
+    try:
+        value = int(seeded) if seeded is not None else None
+    except Exception:
+        value = None
+    if value in SUPPORTED_TEAM_COUNTS:
+        return value
+    populated = max(teams_by_number.keys(), default=0)
+    if populated in SUPPORTED_TEAM_COUNTS:
+        return populated
+    return LEGACY_DEFAULT_TEAM_COUNT
+
+
+def _team_rows_for_editor(teams_by_number: dict[int, dict], id_to_name: dict[Any, str], slot_count: int) -> pd.DataFrame:
+    max_slot = max(slot_count, max(teams_by_number.keys(), default=0))
+    rows = []
+    for slot in range(1, max_slot + 1):
+        team = teams_by_number.get(slot) or {}
+        rows.append(
+            {
+                "Team / Slot": slot,
+                "Player 1": id_to_name.get(team.get("player1_id")),
+                "Player 2": id_to_name.get(team.get("player2_id")),
+                "Source": _safe_text(team.get("source") or "MANUAL"),
+                "Notes": _safe_text(team.get("notes")),
+            }
+        )
+    return pd.DataFrame(rows)
+
+
+def _scoped_query(table, tournament_id: str, draw_id: str | None):
+    query = table.eq("tournament_id", tournament_id)
+    if draw_id:
+        return query.eq("draw_id", draw_id)
+    return query.is_("draw_id", "null")
+
+
+def _teams_ready(teams_by_number: dict[int, dict], team_count: int, *, singles_division: bool) -> bool:
+    if team_count <= 0 or len(teams_by_number) != team_count:
+        return False
+    for num in range(1, team_count + 1):
+        team = teams_by_number.get(num)
+        if not team:
+            return False
+        if not team.get("player1_id"):
+            return False
+        if not singles_division and not team.get("player2_id"):
+            return False
+    return True
+
+
+def render(ctx):
+    mode_label = "Public" if bool(ctx.public_mode) else "Admin"
+    page_shell(
+        "📋 Tournament Operations",
+        "Run registrations, draw prep, team management, imports, and pre-live operations for each tournament division.",
+        mode_label=mode_label,
+    )
+
+    if not bool(getattr(ctx, "admin_logged_in", False)):
+        st.error("Admin login required.")
+        st.stop()
+
+    supabase = getattr(ctx, "supabase", None)
+    club_id = getattr(ctx, "club_id", None)
+    df_players_all = getattr(ctx, "df_players_all", None)
+    name_to_id = getattr(ctx, "name_to_id", {})
+    id_to_name = getattr(ctx, "id_to_name", {})
+
+    if supabase is None or club_id is None:
+        st.error("Missing database context.")
+        st.stop()
+
+    if df_players_all is None or df_players_all.empty:
+        st.warning("No players loaded.")
+        st.stop()
+
+    player_names = sorted(df_players_all["name"].dropna().astype(str).tolist())
+
+    st.subheader("Tournament List")
+    show_archived = st.checkbox("Show archived", value=False, key="tournaments_show_archived")
+    st.caption("Archived tournaments are hidden from default selectors and public registration.")
+
+    tournaments = list_existing_tournaments(supabase, str(club_id), include_archived=show_archived)
+
+    if not tournaments:
+        st.info("No tournaments available for this filter.")
+        st.stop()
+
+    preselected = _safe_text(st.query_params.get("tournament_id"))
+    tournament_labels = []
+    for row in tournaments:
+        status = _safe_text(row.get("status") or "DRAFT")
+        status_label = "ARCHIVED" if status.upper() == "ARCHIVED" else status
+        tournament_labels.append(f"{row['name']} ({status_label})")
+    default_index = 0
+    if preselected:
+        for idx, row in enumerate(tournaments):
+            if str(row.get("id")) == preselected:
+                default_index = idx
+                break
+    selected_label = st.selectbox("Select tournament", tournament_labels, index=default_index)
+    tournament = tournaments[tournament_labels.index(selected_label)]
+    tournament_id = tournament["id"]
+    st.query_params["tournament_id"] = tournament_id
+
+    available, _ = registration_feature_available(supabase)
+    registration_bridge = None
+    if available:
+        try:
+            reg_settings = get_registration_settings(supabase, tournament_id, tournament_name=_safe_text(tournament.get("name")))
+            reg_days = list_registration_days(supabase, tournament_id)
+            reg_events = list_registration_event_options(supabase, tournament_id)
+            reg_state = build_registration_state(supabase, tournament, reg_settings, reg_days, reg_events)
+            registration_bridge = {"settings": reg_settings, "days": reg_days, "events": reg_events, "state": reg_state}
+        except Exception:
+            registration_bridge = None
+
+    st.subheader("Tournament Overview")
+    summary_cols = st.columns(4)
+    summary_cols[0].metric("Status", _safe_text(tournament.get("status") or "DRAFT"))
+    summary_cols[1].metric("Start", _safe_text(tournament.get("start_date") or "—"))
+    summary_cols[2].metric("End", _safe_text(tournament.get("end_date") or "—"))
+    summary_cols[3].metric(
+        "Configured divisions",
+        len((registration_bridge or {}).get("events") or []),
+    )
+    date_window = _tournament_date_window_text(tournament)
+    if date_window:
+        st.caption(f"Tournament date window: {date_window}. Tournament Manager can auto-generate day rows from this range.")
+    else:
+        st.info("This tournament has no saved start/end dates yet. Add dates in Tournament Manager to auto-generate the default day schedule.")
+    if st.button("Open Tournament Manager", key=f"open_tournament_manager_{tournament_id}"):
+        _go_to_tournament_manager(tournament_id)
+
+    st.markdown("#### Admin Actions")
+    st.caption("Delete Draft is only available for empty draft shells with no registrations, draws, teams, games, or podium.")
+    action_cols = st.columns(2)
+    if _safe_text(tournament.get("status")).upper() == "ARCHIVED":
+        action_cols[0].caption("Unarchive restores this tournament as DRAFT; you can change status afterward if needed.")
+        if action_cols[0].button("Unarchive Tournament", key=f"unarchive_tournament_{tournament_id}"):
+            unarchive_tournament(supabase, tournament_id)
+            st.success("Tournament unarchived and restored to DRAFT.")
+            st.rerun()
+    else:
+        if action_cols[0].button("Archive Tournament", key=f"archive_tournament_{tournament_id}"):
+            archive_tournament(supabase, tournament_id)
+            st.success("Tournament archived and hidden from default selectors and public registration.")
+            st.rerun()
+
+    can_delete, usage_summary, delete_reason = tournament_can_be_deleted(supabase, tournament)
+    if can_delete:
+        delete_confirm = action_cols[1].text_input("Type DELETE to confirm draft deletion", key=f"delete_draft_confirm_{tournament_id}")
+        if action_cols[1].button("Delete Draft", key=f"delete_draft_tournament_{tournament_id}"):
+            if delete_confirm != "DELETE":
+                st.error("Type DELETE exactly to confirm draft deletion.")
+            else:
+                delete_unused_draft_tournament(supabase, tournament)
+                st.success("Draft tournament shell deleted.")
+                st.query_params.pop("tournament_id", None)
+                st.rerun()
+    else:
+        usage_bits = [f"{label}: {usage_summary.get(key, 0)}" for key, label in [("registrations", "registrations"), ("registration_selections", "selections"), ("event_draws", "draws"), ("teams", "teams"), ("games", "games"), ("podium", "podium")]]
+        action_cols[1].warning(delete_reason or "This tournament has existing operational/history records. Archive it instead of deleting.")
+        action_cols[1].caption("Usage summary — " + ", ".join(usage_bits))
+
+    draws = _list_draws(supabase, tournament_id)
+    modern_mode = bool((registration_bridge or {}).get("events"))
+    selected_draw: dict[str, Any] | None = None
+    selected_draw_id: str | None = None
+    selected_event_option_id: str | None = None
+    selected_day_id: str | None = None
+    selected_event_type = ""
+    selected_event: dict[str, Any] | None = None
+
+    st.subheader("Division Draw Operations")
+    if modern_mode:
+        event_options = (registration_bridge or {}).get("events") or []
+        day_map = {str(d.get("id")): d for d in ((registration_bridge or {}).get("days") or [])}
+        event_labels = []
+        for row in event_options:
+            day = day_map.get(str(row.get("registration_day_id")), {})
+            day_label = _safe_text(day.get("label") or "Day")
+            family = _safe_text(row.get("event_family_label") or row.get("label") or "Event")
+            division = _safe_text(row.get("division_name") or row.get("label") or "Division")
+            event_labels.append(f"{day_label} • {family} • {division}")
+        selected_event_label = st.selectbox("Select a scheduled division", event_labels, key=f"ops_event_{tournament_id}")
+        selected_event = event_options[event_labels.index(selected_event_label)]
+        selected_event_option_id = str(selected_event.get("id"))
+        selected_day_id = str(selected_event.get("registration_day_id"))
+        selected_event_type = _safe_text(selected_event.get("event_type"))
+
+        info_cols = st.columns(3)
+        info_cols[0].metric("Format", _safe_text(selected_event.get("event_format_override") or selected_event.get("event_format_default") or "—").replace("_", " "))
+        info_cols[1].metric("Scoring", _safe_text(selected_event.get("scoring_override") or selected_event.get("scoring_default") or "—").replace("_", " "))
+        info_cols[2].metric("Capacity", _safe_text(selected_event.get("capacity_teams") or "Open"))
+
+        scoped_draws = [
+            row
+            for row in draws
+            if str(row.get("event_option_id")) == selected_event_option_id and str(row.get("registration_day_id")) == selected_day_id
+        ]
+        if scoped_draws:
+            draw_labels = [f"{_safe_text(row.get('name') or 'Draw')} ({_safe_text(row.get('status') or 'draft')})" for row in scoped_draws]
+            chosen_draw = st.selectbox("Select division draw", draw_labels, key=f"ops_draw_{tournament_id}")
+            selected_draw = scoped_draws[draw_labels.index(chosen_draw)]
+            selected_draw_id = str(selected_draw.get("id"))
+
+        action_cols = st.columns(3)
+        if action_cols[0].button("Create draw for this division", key=f"create_draw_{tournament_id}"):
+            _create_draw(
+                supabase,
+                {
+                    "tournament_id": tournament_id,
+                    "registration_day_id": selected_day_id,
+                    "event_option_id": selected_event_option_id,
+                    "name": f"{_safe_text(selected_event.get('division_name') or selected_event.get('label') or 'Division')} Ops Draw",
+                    "status": "draft",
+                },
+            )
+            st.success("Division draw created.")
+            st.rerun()
+        if action_cols[1].button("Configure registration", key=f"goto_reg_{tournament_id}"):
+            st.query_params["page"] = "tournament_manager"
+            st.query_params["tournament_id"] = tournament_id
+            st.rerun()
+        if action_cols[2].button("Refresh draw list", key=f"refresh_draws_{tournament_id}"):
+            st.rerun()
+    else:
+        st.info("This tournament has not been configured in Tournament Manager yet. Create days, events, and divisions there first.")
+
+    _render_registration_bridge(tournament, registration_bridge)
+
+    if modern_mode and not selected_draw_id:
+        st.warning("Create or select a division draw before editing teams or generating brackets for this scheduled division.")
+
+    team_query = _scoped_query(supabase.table("tournament_teams").select("*"), tournament_id, selected_draw_id)
+    game_query = _scoped_query(supabase.table("tournament_games").select("*"), tournament_id, selected_draw_id)
+    teams_resp = team_query.order("team_number").execute()
+    teams = teams_resp.data or []
+    teams_by_number = {int(row["team_number"]): row for row in teams if row.get("team_number") is not None}
+    teams_by_id = {row["id"]: row for row in teams}
+
+    games_resp = game_query.order("rr_round_number", desc=False).order("rr_slot_number", desc=False).execute()
+    games = games_resp.data or []
+    rr_games = [row for row in games if row.get("stage") == "ROUND_ROBIN"]
+    playoff_games = [row for row in games if row.get("stage") == "PLAYOFF"]
+
+    podium_resp = (
+        supabase.table("tournament_podium")
+        .select("*")
+        .eq("tournament_id", tournament_id)
+        .order("placement", desc=False)
+        .execute()
+    )
+    podium_rows = podium_resp.data or []
+    is_complete = tournament.get("status") == "COMPLETE"
+
+    if modern_mode:
+        st.caption("Operations below are scoped to the selected division draw whenever a draw is selected. Legacy tournaments without scheduled divisions still work tournament-wide.")
+
+    ops_tabs = st.tabs(["Teams"])
+    draw_team_count = _effective_team_count(selected_draw, tournament, teams_by_number)
+    singles_division = _safe_text(selected_event_type).upper() == "SINGLES"
+
+    with ops_tabs[0]:
+        st.subheader("Teams")
+        if modern_mode and not selected_draw_id:
+            st.info("Create or select a division draw above to manage teams for this division.")
+        else:
+            st.caption(
+                "Manual team editing and bulk import happen here. This is also where you move confirmed registrations into an operations-ready draw. Current scheduling engine supports these draw sizes: "
+                + ", ".join(str(x) for x in SUPPORTED_TEAM_COUNTS)
+                + "."
+            )
+            ops_cols = st.columns([2, 1, 1])
+            with ops_cols[0]:
+                roster = _find_event_roster(
+                    registration_bridge,
+                    event_option_id=_safe_text(selected_event_option_id),
+                    registration_day_id=_safe_text(selected_day_id),
+                ) if modern_mode else None
+                confirmed_count = len([row for row in (roster or {}).get("entries", []) if _safe_text(row.get("status")) == "CONFIRMED"])
+                st.metric("Confirmed registration entries", confirmed_count)
+            with ops_cols[1]:
+                draw_size = st.selectbox(
+                    "Draw size / team slots",
+                    SUPPORTED_TEAM_COUNTS,
+                    index=SUPPORTED_TEAM_COUNTS.index(draw_team_count) if draw_team_count in SUPPORTED_TEAM_COUNTS else 0,
+                    disabled=bool(games) or is_complete,
+                    key=f"draw_size_{tournament_id}_{selected_draw_id or 'legacy'}",
+                )
+            with ops_cols[2]:
+                if st.button("Save draw size", disabled=bool(games) or is_complete):
+                    _update_ops_field(supabase, tournament_id, selected_draw_id, "team_count", int(draw_size))
+                    st.success("Draw size updated.")
+                    st.rerun()
+
+            build_cols = st.columns(2)
+            with build_cols[0]:
+                import_mode = st.radio(
+                    "Registration import mode",
+                    ["Append", "Replace"],
+                    horizontal=True,
+                    key=f"reg_import_mode_{tournament_id}_{selected_draw_id or 'legacy'}",
+                )
+            with build_cols[1]:
+                if modern_mode and selected_draw_id and st.button("Build draw from confirmed registrations"):
+                    roster = _find_event_roster(
+                        registration_bridge,
+                        event_option_id=_safe_text(selected_event_option_id),
+                        registration_day_id=_safe_text(selected_day_id),
+                    ) or {}
+                    roster_entries = roster.get("entries") or []
+                    entries = [row for row in roster_entries if _safe_text(row.get("status")) == "CONFIRMED"]
+                    if not entries:
+                        unresolved_partner_count = sum(
+                            1
+                            for row in roster_entries
+                            if _safe_text(row.get("status")).upper() in {"REVIEW", "NEEDS_PARTNER", "PARTNER_MISSING"}
+                        )
+                        if not roster_entries:
+                            st.info("No registration entries exist for this division yet.")
+                        elif unresolved_partner_count:
+                            st.warning(
+                                "No confirmed roster entries are ready yet because partner/roster issues still need admin review."
+                            )
+                            st.caption("Hint: Open Tournament Operations → Registration Admin Review to resolve statuses.")
+                        else:
+                            st.info("No confirmed roster entries are ready for this division.")
+                    else:
+                        if import_mode == "Replace":
+                            _scoped_query(supabase.table("tournament_teams").delete(), tournament_id, selected_draw_id).execute()
+                        start_slot = max(teams_by_number.keys(), default=0) + 1 if import_mode == "Append" else 1
+                        payload = []
+                        unresolved_names = []
+                        for offset, entry in enumerate(entries, start=0):
+                            members = entry.get("members") or []
+                            p1_name = _safe_text((members[0] if len(members) > 0 else {}).get("display_name"))
+                            p2_name = _safe_text((members[1] if len(members) > 1 else {}).get("display_name"))
+                            p1_id = _resolve_player_id(p1_name, name_to_id, id_to_name)
+                            p2_id = _resolve_player_id(p2_name, name_to_id, id_to_name) if p2_name else None
+                            if p1_name and not p1_id:
+                                unresolved_names.append(p1_name)
+                            if p2_name and not p2_id:
+                                unresolved_names.append(p2_name)
+                            payload.append(
+                                {
+                                    "tournament_id": tournament_id,
+                                    "draw_id": selected_draw_id,
+                                    "event_option_id": selected_event_option_id,
+                                    "registration_day_id": selected_day_id,
+                                    "team_number": start_slot + offset,
+                                    "player1_id": p1_id,
+                                    "player2_id": p2_id,
+                                    "source": "REGISTRATION",
+                                    "notes": None,
+                                }
+                            )
+                        if unresolved_names:
+                            st.error("Some confirmed registration names could not be matched to JUPR players: " + ", ".join(sorted(set(unresolved_names))))
+                            st.caption("Hint: fix display names or player mappings first, then rerun this import.")
+                        else:
+                            supabase.table("tournament_teams").upsert(payload, on_conflict="tournament_id,draw_id,team_number").execute()
+                            total_slots = max(draw_size, start_slot + len(payload) - 1)
+                            if total_slots in SUPPORTED_TEAM_COUNTS:
+                                _update_ops_field(supabase, tournament_id, selected_draw_id, "team_count", total_slots)
+                            st.success(f"Imported {len(payload)} confirmed entries into the division draw.")
+                            st.rerun()
+
+            editor_df = st.data_editor(
+                _team_rows_for_editor(teams_by_number, id_to_name, int(draw_size)),
+                use_container_width=True,
+                hide_index=True,
+                key=f"tourney_team_editor_{tournament_id}_{selected_draw_id or 'legacy'}",
+                disabled=is_complete,
+                column_config={
+                    "Team / Slot": st.column_config.NumberColumn("Team / Slot", step=1, min_value=1),
+                    "Player 1": st.column_config.SelectboxColumn("Player 1", options=player_names, required=True),
+                    "Player 2": st.column_config.SelectboxColumn("Player 2", options=player_names, required=not singles_division),
+                    "Source": st.column_config.SelectboxColumn("Source", options=IMPORT_SOURCE_OPTIONS),
+                    "Notes": st.column_config.TextColumn("Notes"),
+                },
+            )
+
+            save_col, import_col = st.columns(2)
+            with save_col:
+                if st.button("Save teams", disabled=is_complete):
+                    seen_slots: set[int] = set()
+                    selected_ids: list[Any] = []
+                    payload = []
+                    for _, row in editor_df.iterrows():
+                        slot = _safe_text(row.get("Team / Slot"))
+                        if not slot:
+                            continue
+                        try:
+                            slot_num = int(float(slot))
+                        except Exception:
+                            st.error("Each row needs a numeric Team / Slot value.")
+                            st.stop()
+                        if slot_num in seen_slots:
+                            st.error(f"Duplicate slot detected: {slot_num}.")
+                            st.stop()
+                        seen_slots.add(slot_num)
+                        p1 = _safe_text(row.get("Player 1"))
+                        p2 = _safe_text(row.get("Player 2"))
+                        if not p1 and not p2:
+                            continue
+                        if singles_division and not p1:
+                            st.error("Singles divisions require Player 1.")
+                            st.stop()
+                        if not singles_division and (not p1 or not p2):
+                            st.error("Doubles divisions require both Player 1 and Player 2.")
+                            st.stop()
+                        if p1 and p2 and p1 == p2:
+                            st.error("A team cannot use the same player twice.")
+                            st.stop()
+                        p1_id = _resolve_player_id(p1, name_to_id, id_to_name)
+                        p2_id = _resolve_player_id(p2, name_to_id, id_to_name) if p2 else None
+                        if p1_id:
+                            selected_ids.append(p1_id)
+                        if p2_id:
+                            selected_ids.append(p2_id)
+                        payload.append(
+                            {
+                                "tournament_id": tournament_id,
+                                "draw_id": selected_draw_id,
+                                "event_option_id": selected_event_option_id,
+                                "registration_day_id": selected_day_id,
+                                "team_number": slot_num,
+                                "player1_id": p1_id,
+                                "player2_id": p2_id,
+                                "source": _safe_text(row.get("Source") or "MANUAL"),
+                                "notes": _safe_text(row.get("Notes")) or None,
+                            }
+                        )
+
+                    duplicates = {pid for pid in selected_ids if pid is not None and selected_ids.count(pid) > 1}
+                    if duplicates:
+                        names = ", ".join(id_to_name.get(pid, str(pid)) for pid in duplicates)
+                        st.error(f"Duplicate players detected in this draw: {names}.")
+                        st.stop()
+                    _scoped_query(supabase.table("tournament_teams").delete(), tournament_id, selected_draw_id).execute()
+                    if payload:
+                        supabase.table("tournament_teams").upsert(payload, on_conflict="tournament_id,draw_id,team_number").execute()
+                    st.success("Teams saved.")
+                    st.rerun()
+
+            with import_col:
+                with st.expander("Bulk upload teams"):
+                    upload_mode = st.radio(
+                        "Bulk upload save mode",
+                        ["Append", "Replace"],
+                        horizontal=True,
+                        key=f"bulk_mode_{tournament_id}_{selected_draw_id or 'legacy'}",
+                    )
+                    uploaded = st.file_uploader("CSV or XLSX", type=["csv", "xlsx"], key=f"bulk_file_{tournament_id}_{selected_draw_id or 'legacy'}")
+                    pasted = st.text_area("Or paste CSV/TSV", key=f"bulk_text_{tournament_id}_{selected_draw_id or 'legacy'}")
+                    if st.button("Preview and save bulk upload", key=f"bulk_preview_{tournament_id}_{selected_draw_id or 'legacy'}"):
+                        try:
+                            parsed = _canonical_import_df(_parse_bulk_upload(uploaded, pasted))
+                        except Exception as exc:
+                            st.error(f"Could not parse upload: {exc}")
+                            st.stop()
+                        if parsed.empty:
+                            st.warning("No rows found.")
+                            st.stop()
+                        parsed = parsed.fillna("")
+                        unresolved = []
+                        duplicate_names = []
+                        id_counts: dict[Any, int] = {}
+                        preview_rows = []
+                        for _, row in parsed.iterrows():
+                            p1_name = _safe_text(row.get("Player 1"))
+                            p2_name = _safe_text(row.get("Player 2"))
+                            p1_id = _resolve_player_id(p1_name, name_to_id, id_to_name)
+                            p2_id = _resolve_player_id(p2_name, name_to_id, id_to_name) if p2_name else None
+                            if not p1_id:
+                                unresolved.append(p1_name)
+                            if p2_name and not p2_id:
+                                unresolved.append(p2_name)
+                            for pid in [p1_id, p2_id]:
+                                if pid:
+                                    id_counts[pid] = id_counts.get(pid, 0) + 1
+                            preview_rows.append(
+                                {
+                                    "Player 1": p1_name,
+                                    "Player 2": p2_name,
+                                    "Player 1 Resolved": bool(p1_id),
+                                    "Player 2 Resolved": True if not p2_name else bool(p2_id),
+                                    "Seed": row.get("Seed"),
+                                    "Notes": _safe_text(row.get("Notes")),
+                                }
+                            )
+                        duplicate_names = [id_to_name.get(pid, str(pid)) for pid, count in id_counts.items() if count > 1]
+                        st.dataframe(pd.DataFrame(preview_rows), use_container_width=True, hide_index=True)
+                        if unresolved:
+                            st.error("Unresolved player names: " + ", ".join(sorted({name for name in unresolved if name})))
+                        if duplicate_names:
+                            st.error("Duplicate players in upload: " + ", ".join(sorted(set(duplicate_names))))
+                        if unresolved or duplicate_names:
+                            st.stop()
+
+                        if upload_mode == "Replace":
+                            _scoped_query(supabase.table("tournament_teams").delete(), tournament_id, selected_draw_id).execute()
+                            start_slot = 1
+                        else:
+                            start_slot = max(teams_by_number.keys(), default=0) + 1
+
+                        payload = []
+                        for offset, row in enumerate(preview_rows, start=0):
+                            payload.append(
+                                {
+                                    "tournament_id": tournament_id,
+                                    "draw_id": selected_draw_id,
+                                    "event_option_id": selected_event_option_id,
+                                    "registration_day_id": selected_day_id,
+                                    "team_number": start_slot + offset,
+                                    "player1_id": _resolve_player_id(_safe_text(row.get("Player 1")), name_to_id, id_to_name),
+                                    "player2_id": _resolve_player_id(_safe_text(row.get("Player 2")), name_to_id, id_to_name) if _safe_text(row.get("Player 2")) else None,
+                                    "seed": int(float(row.get("Seed"))) if _safe_text(row.get("Seed")).replace(".", "", 1).isdigit() else None,
+                                    "source": "BULK_UPLOAD",
+                                    "notes": _safe_text(row.get("Notes")) or None,
+                                }
+                            )
+                        supabase.table("tournament_teams").upsert(payload, on_conflict="tournament_id,draw_id,team_number").execute()
+                        final_slot = start_slot + len(payload) - 1
+                        if final_slot in SUPPORTED_TEAM_COUNTS:
+                            _update_ops_field(supabase, tournament_id, selected_draw_id, "team_count", final_slot)
+                        st.success("Bulk upload saved.")
+                        st.rerun()
+
+            with st.expander("Import Results into This Draw"):
+                st.caption(
+                    f"Selected tournament: {_safe_text(tournament.get('name'))} • division: {_safe_text((selected_event or {}).get('division_name') or (selected_event or {}).get('label') or '—')} • draw: {_safe_text((selected_draw or {}).get('name') or selected_draw_id)}. "
+                    "This workflow writes teams and matches to this draw scope only."
+                )
+                import_scope_key = f"{tournament_id}_{selected_draw_id or 'legacy'}"
+                bundle_state_key = f"draw_results_bundle_{import_scope_key}"
+                mapping_state_key = f"draw_results_mapping_{import_scope_key}"
+                review_state_key = f"draw_results_match_review_{import_scope_key}"
+                mode_state_key = f"draw_results_mode_{import_scope_key}"
+                podium_state_key = f"draw_results_podium_{import_scope_key}"
+
+                upload_file = st.file_uploader(
+                    "Upload DUPR-style results CSV",
+                    type=["csv"],
+                    key=f"draw_results_csv_{import_scope_key}",
+                )
+                if st.button("Parse uploaded results", key=f"draw_results_parse_{import_scope_key}", disabled=upload_file is None):
+                    if upload_file is None:
+                        st.error("Upload a CSV file first.")
+                    else:
+                        parsed_bundle = parse_dupr_results_csv(upload_file.getvalue())
+                        existing_players = []
+                        for _, prow in df_players_all.iterrows():
+                            existing_players.append(
+                                {
+                                    "id": prow.get("id"),
+                                    "name": _safe_text(prow.get("name")),
+                                    "dupr_id": prow.get("dupr_id") or prow.get("duprId") or prow.get("dupr"),
+                                }
+                            )
+                        suggestions = suggest_player_matches(parsed_bundle.get("players") or [], existing_players)
+                        initial_mapping = {}
+                        for imported in parsed_bundle.get("players") or []:
+                            import_key = str(imported.get("import_key"))
+                            suggested = suggestions.get(import_key) or {}
+                            if suggested.get("suggested_player_id"):
+                                initial_mapping[import_key] = {
+                                    "action": "use_existing",
+                                    "player_id": suggested.get("suggested_player_id"),
+                                }
+                            else:
+                                initial_mapping[import_key] = {"action": "unresolved", "player_id": None}
+                        initial_reviews = {
+                            str(row.get("source_row")): {"include": True, "stage": "PLAYOFF"}
+                            for row in (parsed_bundle.get("matches") or [])
+                        }
+                        st.session_state[bundle_state_key] = {**parsed_bundle, "suggestions": suggestions}
+                        st.session_state[mapping_state_key] = initial_mapping
+                        st.session_state[review_state_key] = initial_reviews
+                        st.session_state[mode_state_key] = "Replace"
+                        st.session_state[podium_state_key] = {}
+
+                bundle = st.session_state.get(bundle_state_key)
+                if not bundle:
+                    st.info("Upload and parse a DUPR-style CSV to start the draw-scoped import wizard.")
+                else:
+                    for err in bundle.get("errors") or []:
+                        st.error(err)
+                    for warn in bundle.get("warnings") or []:
+                        st.warning(warn)
+                    if bundle.get("errors"):
+                        st.stop()
+
+                    st.markdown("#### 1) Player mapping review")
+                    mapping_decisions = st.session_state.get(mapping_state_key, {})
+                    suggestions = bundle.get("suggestions") or {}
+                    duplicate_existing_ids: dict[str, int] = {}
+                    for imported in bundle.get("players") or []:
+                        import_key = str(imported.get("import_key"))
+                        suggested = suggestions.get(import_key) or {}
+                        current = mapping_decisions.get(import_key) or {"action": "unresolved", "player_id": None}
+                        with st.container(border=True):
+                            cols = st.columns([3, 2, 3])
+                            cols[0].markdown(
+                                f"**{_safe_text(imported.get('display_name') or 'Unknown')}**  \nDUPR ID: `{_safe_text(imported.get('dupr_id') or '—')}`"
+                            )
+                            cols[1].caption(
+                                "Suggested match: "
+                                + (
+                                    f"{id_to_name.get(suggested.get('suggested_player_id'), suggested.get('suggested_player_id'))} ({_safe_text(suggested.get('reason'))})"
+                                    if suggested.get("suggested_player_id")
+                                    else _safe_text(suggested.get("reason") or "none")
+                                )
+                            )
+                            action_choice = cols[2].selectbox(
+                                "Action",
+                                ["unresolved", "use_existing", "create_new"],
+                                index=["unresolved", "use_existing", "create_new"].index(
+                                    current.get("action") if current.get("action") in {"unresolved", "use_existing", "create_new"} else "unresolved"
+                                ),
+                                format_func=lambda v: {"unresolved": "Unresolved", "use_existing": "Use existing player", "create_new": "Create new player"}[v],
+                                key=f"map_action_{import_scope_key}_{import_key}",
+                            )
+                            mapped_id = current.get("player_id")
+                            if action_choice == "use_existing":
+                                options = [None, *sorted(id_to_name.keys(), key=lambda pid: _safe_text(id_to_name.get(pid)))]
+                                default_idx = options.index(mapped_id) if mapped_id in options else (options.index(suggested.get("suggested_player_id")) if suggested.get("suggested_player_id") in options else 0)
+                                mapped_id = st.selectbox(
+                                    "Choose existing JUPR player",
+                                    options,
+                                    index=default_idx,
+                                    format_func=lambda pid: id_to_name.get(pid, "Select player") if pid else "Select player",
+                                    key=f"map_existing_{import_scope_key}_{import_key}",
+                                )
+                            else:
+                                mapped_id = None
+                            mapping_decisions[import_key] = {"action": action_choice, "player_id": mapped_id}
+                            if action_choice == "use_existing" and mapped_id is not None:
+                                dup_key = str(mapped_id)
+                                duplicate_existing_ids[dup_key] = duplicate_existing_ids.get(dup_key, 0) + 1
+
+                    st.session_state[mapping_state_key] = mapping_decisions
+                    duplicate_warning_ids = [pid for pid, count in duplicate_existing_ids.items() if count > 1]
+                    allow_duplicate_mapping = st.checkbox(
+                        "Allow duplicate mapping of multiple imported players to the same existing JUPR player (not recommended)",
+                        value=False,
+                        key=f"allow_dup_map_{import_scope_key}",
+                    )
+                    if duplicate_warning_ids:
+                        duplicate_names = ", ".join(_safe_text(id_to_name.get(pid, pid)) for pid in duplicate_warning_ids)
+                        st.warning(f"Duplicate existing-player mappings detected: {duplicate_names}.")
+
+                    st.markdown("#### 2) Match review and stage assignment")
+                    review_map = st.session_state.get(review_state_key, {})
+                    match_rows = []
+                    for row in bundle.get("matches") or []:
+                        source_row = str(row.get("source_row"))
+                        review = review_map.get(source_row, {"include": True, "stage": "PLAYOFF"})
+                        team_a = row.get("team_a_key")
+                        team_b = row.get("team_b_key")
+                        match_rows.append(
+                            {
+                                "source_row": int(row.get("source_row") or 0),
+                                "team_a": team_a,
+                                "team_b": team_b,
+                                "score_summary": _safe_text(row.get("score_summary") or "—"),
+                                "winner": _safe_text(row.get("winner_side") or "—"),
+                                "stage": review.get("stage") or "PLAYOFF",
+                                "include": bool(review.get("include", True)),
+                            }
+                        )
+                    review_df = st.data_editor(
+                        pd.DataFrame(match_rows),
+                        key=f"draw_results_match_editor_{import_scope_key}",
+                        hide_index=True,
+                        use_container_width=True,
+                        column_config={
+                            "stage": st.column_config.SelectboxColumn("stage", options=STAGE_OPTIONS, required=True),
+                            "include": st.column_config.CheckboxColumn("include"),
+                        },
+                    )
+                    bulk_cols = st.columns(2)
+                    if bulk_cols[0].button("Set all to PLAYOFF", key=f"draw_results_bulk_playoff_{import_scope_key}"):
+                        review_df["stage"] = "PLAYOFF"
+                    if bulk_cols[1].button("Set all to ROUND_ROBIN", key=f"draw_results_bulk_rr_{import_scope_key}"):
+                        review_df["stage"] = "ROUND_ROBIN"
+                    st.session_state[review_state_key] = {
+                        str(int(row["source_row"])): {"include": bool(row.get("include", True)), "stage": _safe_text(row.get("stage") or "PLAYOFF")}
+                        for _, row in review_df.iterrows()
+                    }
+
+                    import_mode = st.radio(
+                        "Import mode",
+                        ["Replace", "Append"],
+                        horizontal=True,
+                        index=0 if st.session_state.get(mode_state_key, "Replace") == "Replace" else 1,
+                        key=f"draw_results_mode_radio_{import_scope_key}",
+                    )
+                    st.session_state[mode_state_key] = import_mode
+                    payload_preview = build_draw_import_payload(
+                        bundle=bundle,
+                        mapping_decisions=mapping_decisions,
+                        match_reviews=st.session_state.get(review_state_key, {}),
+                    )
+                    for warn in payload_preview.get("warnings") or []:
+                        st.warning(warn)
+                    for err in payload_preview.get("errors") or []:
+                        st.error(err)
+
+                    st.markdown("#### 3) Podium review (tournament context)")
+                    if modern_mode:
+                        st.caption("Podium save uses current tournament podium infrastructure; this section reflects the selected draw import review.")
+                    team_ref_options = [row.get("team_ref") for row in (payload_preview.get("teams") or [])]
+                    team_ref_labels = {}
+                    for team_row in payload_preview.get("teams") or []:
+                        p1 = _safe_text(team_row.get("p1_ref"))
+                        p2 = _safe_text(team_row.get("p2_ref"))
+                        team_ref_labels[team_row.get("team_ref")] = f"{p1} / {p2}" if p2 else p1
+                    defaults = payload_preview.get("podium_candidates") or []
+                    podium_choices = st.session_state.get(podium_state_key, {})
+                    podium_choices["1"] = st.selectbox("1st place", [None, *team_ref_options], index=([None, *team_ref_options].index(podium_choices.get("1")) if podium_choices.get("1") in [None, *team_ref_options] else ([None, *team_ref_options].index(defaults[0]) if len(defaults) > 0 and defaults[0] in [None, *team_ref_options] else 0)), format_func=lambda value: team_ref_labels.get(value, "Select team"), key=f"draw_import_podium_1_{import_scope_key}")
+                    podium_choices["2"] = st.selectbox("2nd place", [None, *team_ref_options], index=([None, *team_ref_options].index(podium_choices.get("2")) if podium_choices.get("2") in [None, *team_ref_options] else ([None, *team_ref_options].index(defaults[1]) if len(defaults) > 1 and defaults[1] in [None, *team_ref_options] else 0)), format_func=lambda value: team_ref_labels.get(value, "Select team"), key=f"draw_import_podium_2_{import_scope_key}")
+                    podium_choices["3"] = st.selectbox("3rd place", [None, *team_ref_options], index=([None, *team_ref_options].index(podium_choices.get("3")) if podium_choices.get("3") in [None, *team_ref_options] else ([None, *team_ref_options].index(defaults[2]) if len(defaults) > 2 and defaults[2] in [None, *team_ref_options] else 0)), format_func=lambda value: team_ref_labels.get(value, "Select team"), key=f"draw_import_podium_3_{import_scope_key}")
+                    st.session_state[podium_state_key] = podium_choices
+
+                    with st.form(key=f"draw_results_commit_form_{import_scope_key}"):
+                        st.markdown("#### 4) Final commit")
+                        replace_confirm = st.text_input("If Replace mode is selected, type REPLACE to confirm scoped delete.")
+                        submit_import = st.form_submit_button("Commit import into selected draw", type="primary")
+                        if submit_import:
+                            if duplicate_warning_ids and not allow_duplicate_mapping:
+                                st.error("Duplicate existing-player mappings are present. Either resolve them or explicitly allow them.")
+                                st.stop()
+                            final_payload = build_draw_import_payload(
+                                bundle=bundle,
+                                mapping_decisions=st.session_state.get(mapping_state_key, {}),
+                                match_reviews=st.session_state.get(review_state_key, {}),
+                            )
+                            if final_payload.get("errors"):
+                                st.error("Import blocked due to validation errors.")
+                                st.stop()
+                            if import_mode == "Replace" and replace_confirm.strip().upper() != "REPLACE":
+                                st.error("Replace mode requires typing REPLACE for confirmation.")
+                                st.stop()
+
+                            players_resp = supabase.table("players").select("id,name").eq("club_id", str(club_id)).execute()
+                            players_rows = players_resp.data or []
+                            players_by_norm: dict[str, Any] = {}
+                            for prow in players_rows:
+                                norm_name = _normalize_name(prow.get("name"))
+                                if norm_name and norm_name not in players_by_norm:
+                                    players_by_norm[norm_name] = prow.get("id")
+
+                            player_id_by_ref: dict[str, Any] = {}
+                            imported_by_key = {str(row.get("import_key")): row for row in (bundle.get("players") or [])}
+                            for import_key, ref in (final_payload.get("mapped_player_refs") or {}).items():
+                                if str(ref).startswith("existing:"):
+                                    player_id_by_ref[ref] = str(ref).split("existing:", 1)[1]
+                                    continue
+                                imported = imported_by_key.get(import_key) or {}
+                                display_name = _safe_text(imported.get("display_name"))
+                                norm_name = _normalize_name(display_name)
+                                existing_id = players_by_norm.get(norm_name)
+                                if existing_id:
+                                    player_id_by_ref[ref] = existing_id
+                                    continue
+                                ok, err = safe_add_player(
+                                    supabase=supabase,
+                                    club_id=str(club_id),
+                                    name=display_name,
+                                    rating_jupr=10.0,
+                                )
+                                if not ok:
+                                    st.error(f"Failed creating player {display_name}: {err}")
+                                    st.stop()
+                                refresh_resp = supabase.table("players").select("id,name").eq("club_id", str(club_id)).execute()
+                                refresh_rows = refresh_resp.data or []
+                                for row in refresh_rows:
+                                    norm = _normalize_name(row.get("name"))
+                                    if norm and norm not in players_by_norm:
+                                        players_by_norm[norm] = row.get("id")
+                                if norm_name not in players_by_norm:
+                                    st.error(f"Could not resolve player id after create: {display_name}")
+                                    st.stop()
+                                player_id_by_ref[ref] = players_by_norm[norm_name]
+
+                            if import_mode == "Replace":
+                                _scoped_query(supabase.table("tournament_games").delete(), tournament_id, selected_draw_id).execute()
+                                _scoped_query(supabase.table("tournament_teams").delete(), tournament_id, selected_draw_id).execute()
+                                team_start_slot = 1
+                            else:
+                                team_start_slot = max(teams_by_number.keys(), default=0) + 1
+
+                            team_rows_to_write = []
+                            team_ref_to_number: dict[str, int] = {}
+                            for idx, team_row in enumerate(final_payload.get("teams") or [], start=0):
+                                team_number = team_start_slot + idx
+                                team_ref_to_number[team_row.get("team_ref")] = team_number
+                                team_rows_to_write.append(
+                                    {
+                                        "tournament_id": tournament_id,
+                                        "draw_id": selected_draw_id,
+                                        "event_option_id": selected_event_option_id,
+                                        "registration_day_id": selected_day_id,
+                                        "team_number": team_number,
+                                        "player1_id": player_id_by_ref.get(team_row.get("p1_ref")),
+                                        "player2_id": player_id_by_ref.get(team_row.get("p2_ref")) if team_row.get("p2_ref") else None,
+                                        "source": "BULK_UPLOAD",
+                                        "notes": "RESULTS_IMPORT",
+                                    }
+                                )
+                            if not team_rows_to_write:
+                                st.error("No teams to import.")
+                                st.stop()
+                            supabase.table("tournament_teams").upsert(team_rows_to_write, on_conflict="tournament_id,draw_id,team_number").execute()
+
+                            refreshed_teams = _scoped_query(supabase.table("tournament_teams").select("*"), tournament_id, selected_draw_id).order("team_number").execute().data or []
+                            team_id_by_number = {int(row.get("team_number")): row.get("id") for row in refreshed_teams if row.get("team_number") is not None}
+
+                            game_rows_to_write = []
+                            for match_row in final_payload.get("matches") or []:
+                                team_a_num = team_ref_to_number.get(match_row.get("team_a_ref"))
+                                team_b_num = team_ref_to_number.get(match_row.get("team_b_ref"))
+                                team_a_id = team_id_by_number.get(team_a_num) if team_a_num is not None else None
+                                team_b_id = team_id_by_number.get(team_b_num) if team_b_num is not None else None
+                                if not team_a_id or not team_b_id:
+                                    st.error(f"Unwriteable match payload for source row {match_row.get('source_row')}.")
+                                    st.stop()
+                                game_payload = {
+                                    "tournament_id": tournament_id,
+                                    "draw_id": selected_draw_id,
+                                    "event_option_id": selected_event_option_id,
+                                    "registration_day_id": selected_day_id,
+                                    "team_a_id": team_a_id,
+                                    "team_b_id": team_b_id,
+                                    "stage": match_row.get("stage") or "PLAYOFF",
+                                    "score_a": match_row.get("score_a"),
+                                    "score_b": match_row.get("score_b"),
+                                }
+                                if match_row.get("score_a") is not None and match_row.get("score_b") is not None:
+                                    game_payload.update(finalize_game(game_payload))
+                                game_rows_to_write.append(game_payload)
+                            if game_rows_to_write:
+                                supabase.table("tournament_games").insert(game_rows_to_write).execute()
+
+                            team_id_by_ref: dict[str, Any] = {}
+                            for team_ref, team_number in team_ref_to_number.items():
+                                team_id_by_ref[team_ref] = team_id_by_number.get(team_number)
+                            podium_refs = st.session_state.get(podium_state_key, {})
+                            placements = []
+                            for placement in [1, 2, 3]:
+                                team_ref = podium_refs.get(str(placement))
+                                team_id = team_id_by_ref.get(team_ref)
+                                if team_id:
+                                    placements.append({"placement": placement, "team_id": team_id})
+                            if placements:
+                                podium_payload = build_podium_payload(tournament_id, placements, source="DRAW_RESULTS_IMPORT")
+                                upsert_tournament_podium(supabase, tournament_id, podium_payload)
+
+                            st.success(f"Imported {len(team_rows_to_write)} teams and {len(game_rows_to_write)} matches into the selected draw.")
+                            st.rerun()
+
+
+
+
+    st.divider()
+    st.subheader("Registration Admin Review")
+    raw_regs = list_registrations(supabase, tournament_id)
+    if not raw_regs:
+        st.info("No registration submissions to review yet.")
+    else:
+        options = {
+            f"{_safe_text(row.get('display_name') or row.get('email') or row.get('id'))} · {_safe_text(row.get('status'))}/{_safe_text(row.get('payment_status'))} · {row.get('submitted_at') or 'no timestamp'}": row
+            for row in raw_regs
+        }
+        selected_label = st.selectbox("Select registration to review", list(options.keys()), key=f"ops_admin_pick_{tournament_id}")
+        selected_reg = options[selected_label]
+        with st.form(f"ops_admin_review_{tournament_id}"):
+            st.text_input("Display name", value=_safe_text(selected_reg.get("display_name")), disabled=True)
+            st.text_input("Email", value=_safe_text(selected_reg.get("email")), disabled=True)
+            admin_status = st.selectbox(
+                "Registration status",
+                ADMIN_REGISTRATION_STATUS_OPTIONS,
+                index=ADMIN_REGISTRATION_STATUS_OPTIONS.index(_safe_text(selected_reg.get("status")).lower())
+                if _safe_text(selected_reg.get("status")).lower() in ADMIN_REGISTRATION_STATUS_OPTIONS
+                else 0,
+            )
+            admin_payment_status = st.selectbox(
+                "Payment status",
+                ADMIN_PAYMENT_STATUS_OPTIONS,
+                index=ADMIN_PAYMENT_STATUS_OPTIONS.index(_safe_text(selected_reg.get("payment_status")).lower())
+                if _safe_text(selected_reg.get("payment_status")).lower() in ADMIN_PAYMENT_STATUS_OPTIONS
+                else 0,
+            )
+            save_admin = st.form_submit_button("Save registration admin fields", use_container_width=True)
+        if save_admin:
+            try:
+                update_registration_admin_fields(
+                    supabase,
+                    tournament_id=tournament_id,
+                    registration_id=str(selected_reg.get("id")),
+                    status=admin_status,
+                    payment_status=admin_payment_status,
+                )
+                st.success("Registration admin fields updated.")
+                st.rerun()
+            except Exception as exc:
+                st.error(f"Could not update registration: {exc}")
+
+
+def _render_registration_bridge(tournament: dict[str, Any], registration_bridge: dict[str, Any] | None) -> None:
+    st.subheader("Registration Bridge")
+    if not registration_bridge:
+        st.warning("Registration has not been configured for this tournament yet.")
+        st.info("Use Tournament Manager to define days, events, divisions, and public registration links.")
+        if st.button("Configure Registration", key=f"configure_registration_bridge_{tournament.get('id')}"):
+            _go_to_tournament_manager(tournament.get("id"))
+        return
+
+    settings = registration_bridge.get("settings") or {}
+    state = registration_bridge.get("state") or {}
+    summary = state.get("summary") or {}
+    links = build_public_urls(
+        base_url=_safe_text(st.session_state.get("base_url")),
+        tournament_id=str(tournament.get("id")),
+        registration_slug=settings.get("registration_slug"),
+    )
+
+    cols = st.columns(4)
+    cols[0].metric("Registrations", summary.get("total_registrations", 0))
+    cols[1].metric("Confirmed", summary.get("confirmed_entries", 0))
+    cols[2].metric("Needs partner", summary.get("needs_partner_entries", 0))
+    cols[3].metric("Issues", summary.get("issue_count", 0))
+
+    st.caption("Use Tournament Manager to set up registration. Use this page to build operations-ready division draws from those confirmed registrations.")
+    with st.expander("Registration links"):
+        st.text_input("Tournament Manager", value=links["admin_manager"], key=f"ops_reg_admin_{tournament.get('id')}")
+        st.text_input("Public registration", value=links["registration"], key=f"ops_reg_public_{tournament.get('id')}")
+        st.text_input("Public partner board", value=links["partner_board"], key=f"ops_reg_partner_{tournament.get('id')}")
+
+
+def _render_games_table(*, games, teams_by_id, id_to_name, on_save, key_prefix: str, disabled: bool = False):
+    rounds = sorted({int(game.get("rr_round_number", 0)) for game in games})
+    for round_num in rounds:
+        st.markdown(f"#### Round {round_num}")
+        round_games = [game for game in games if int(game.get("rr_round_number", 0)) == round_num]
+        scores = {}
+        with st.form(key=f"{key_prefix}_round_{round_num}"):
+            for game in round_games:
+                team_a = teams_by_id.get(game.get("team_a_id"), {})
+                team_b = teams_by_id.get(game.get("team_b_id"), {})
+                label_a = _team_label(team_a, id_to_name)
+                label_b = _team_label(team_b, id_to_name)
+                slot = game.get("rr_slot_number")
+                c1, c2, c3, c4 = st.columns([4, 1, 1, 2])
+                c1.write(f"Game {slot}: {label_a} vs {label_b}")
+                c2.number_input("Score A", min_value=0, value=int(game.get("score_a") or 0), key=f"{key_prefix}_a_{game['id']}", disabled=disabled)
+                c3.number_input("Score B", min_value=0, value=int(game.get("score_b") or 0), key=f"{key_prefix}_b_{game['id']}", disabled=disabled)
+                c4.caption("Final" if game.get("finalized_at") else "Open")
+                scores[game["id"]] = game
+            if st.form_submit_button("Save scores", disabled=disabled):
+                on_save(scores)
+
+
+def _render_playoff_bracket(*, games, teams_by_id, id_to_name, on_save, disabled: bool = False):
+    round_order = ["QF", "SF", "Final", "Bronze"]
+    for round_name in round_order:
+        round_games = [game for game in games if game.get("playoff_round") == round_name]
+        if not round_games:
+            continue
+        st.markdown(f"#### {round_name}")
+        with st.form(key=f"playoff_{round_name}"):
+            for game in round_games:
+                team_a = teams_by_id.get(game.get("team_a_id"), {})
+                team_b = teams_by_id.get(game.get("team_b_id"), {})
+                label_a = _team_label(team_a, id_to_name)
+                label_b = _team_label(team_b, id_to_name)
+                c1, c2, c3, c4 = st.columns([4, 1, 1, 2])
+                c1.write(f"{game.get('playoff_game_code')}: {label_a} vs {label_b}")
+                c2.number_input("Score A", min_value=0, value=int(game.get("score_a") or 0), key=f"playoff_a_{game['id']}", disabled=disabled or not game.get("team_a_id") or not game.get("team_b_id"))
+                c3.number_input("Score B", min_value=0, value=int(game.get("score_b") or 0), key=f"playoff_b_{game['id']}", disabled=disabled or not game.get("team_a_id") or not game.get("team_b_id"))
+                c4.caption("Final" if game.get("finalized_at") else "Open")
+            if st.form_submit_button("Save scores", disabled=disabled):
+                on_save({game["id"]: game for game in round_games})
+
+
+def _render_podium_review(
+    ctx,
+    tournament: dict[str, Any],
+    *,
+    teams_by_id: dict[str, dict],
+    id_to_name: dict[Any, str],
+    rr_games: list[dict],
+    playoff_games: list[dict],
+) -> None:
+    tournament_id = tournament["id"]
+    tournament_name = tournament.get("name", "Tournament")
+    team_count = len(teams_by_id)
+    max_places = min(3, team_count)
+
+    rr_podium = compute_podium_from_rr(list(teams_by_id.values()), rr_games, max_placements=max_places) if rr_games else []
+    playoff_podium = compute_podium_from_playoffs(playoff_games)
+    default_mode = "PLAYOFF_RESULTS" if playoff_podium else "ROUND_ROBIN_STANDINGS"
+    mode_labels = {"PLAYOFF_RESULTS": "Use Playoff Results", "ROUND_ROBIN_STANDINGS": "Use RR Standings", "MANUAL": "Manual"}
+    mode_options = ["PLAYOFF_RESULTS", "ROUND_ROBIN_STANDINGS", "MANUAL"]
+    mode = st.radio(
+        "Podium review",
+        mode_options,
+        index=mode_options.index(default_mode),
+        format_func=lambda value: mode_labels[value],
+        key=f"podium_review_mode_{tournament_id}",
+    )
+
+    placements: list[dict[str, str | int]] = []
+    source = "MANUAL"
+    if mode == "PLAYOFF_RESULTS":
+        source = "PLAYOFF"
+        if not playoff_podium:
+            st.error("Playoff results are not finalized yet.")
+        else:
+            placements = playoff_podium
+            _render_podium_preview(placements, teams_by_id, id_to_name)
+    elif mode == "ROUND_ROBIN_STANDINGS":
+        source = "ROUND_ROBIN"
+        if not rr_games:
+            st.error("Round robin games are required to use standings.")
+        else:
+            if _has_unscored_rr_games(rr_games):
+                st.warning("Some round robin games are unscored; podium reflects currently entered results.")
+            placements = rr_podium
+            _render_podium_preview(placements, teams_by_id, id_to_name)
+    else:
+        options = list(teams_by_id.keys())
+        for placement in range(1, max_places + 1):
+            selection = st.selectbox(
+                f"{_placement_label(placement)}",
+                options=[None, *options],
+                format_func=lambda value: _team_label(teams_by_id.get(value, {}), id_to_name) if value else "Select team",
+                key=f"podium_manual_{tournament_id}_{placement}",
+            )
+            if selection:
+                placements.append({"placement": placement, "team_id": selection})
+
+    if st.button("Finalize tournament", type="primary", disabled=max_places == 0):
+        if mode == "MANUAL" and len(placements) < max_places:
+            st.error("Select a team for each podium placement.")
+            return
+        try:
+            payload = build_podium_payload(tournament_id, placements, source)
+        except ValueError as exc:
+            st.error(str(exc))
+            return
+        if not payload:
+            st.error("Podium placements are required to complete the tournament.")
+            return
+        upsert_tournament_podium(ctx.supabase, tournament_id, payload)
+        award_tournament_trophies_from_podium(ctx, tournament_id, tournament_name)
+        ctx.supabase.table("tournaments").update({"status": "COMPLETE"}).eq("id", tournament_id).execute()
+        st.success("Tournament completed and podium locked.")
+        st.session_state.pop(f"podium_review_open_{tournament_id}", None)
+        st.rerun()
+
+
+def _render_podium_read_only(podium_rows: list[dict], teams_by_id: dict[str, dict], id_to_name: dict[Any, str]) -> None:
+    if not podium_rows:
+        st.info("No podium recorded yet.")
+        return
+    st.markdown("#### 🏆 Podium")
+    _render_podium_preview(podium_rows, teams_by_id, id_to_name)
+
+
+def _render_podium_preview(placements: list[dict], teams_by_id: dict[str, dict], id_to_name: dict[Any, str]) -> None:
+    rows = []
+    for row in sorted(placements, key=lambda item: int(item.get("placement", 0) or 0)):
+        team = teams_by_id.get(row.get("team_id"), {})
+        rows.append({"Placement": _placement_label(int(row.get("placement", 0) or 0)), "Team": _team_label(team, id_to_name)})
+    if rows:
+        st.dataframe(pd.DataFrame(rows), use_container_width=True, hide_index=True)
+
+
+def _placement_label(placement: int) -> str:
+    medals = {1: "🥇 Champion", 2: "🥈 Runner-up", 3: "🥉 Bronze"}
+    return medals.get(int(placement), f"Placement {placement}")
+
+
+def _has_unscored_rr_games(games: list[dict]) -> bool:
+    for game in games:
+        score_a = game.get("score_a")
+        score_b = game.get("score_b")
+        if score_a is None or score_b is None:
+            return True
+        try:
+            if int(score_a) == 0 and int(score_b) == 0:
+                return True
+        except Exception:
+            return True
+    return False
+
+
+def _team_label(team: dict, id_to_name: dict[Any, str]) -> str:
+    if not team:
+        return "TBD"
+    p1 = id_to_name.get(team.get("player1_id"), "?")
+    p2 = id_to_name.get(team.get("player2_id"))
+    if p2:
+        return f"Team {team.get('team_number')}: {p1} / {p2}"
+    return f"Team {team.get('team_number')}: {p1}"
+
+
+def _save_games(ctx, tournament: dict[str, Any], teams_by_id: dict[str, dict], game_map: dict[str, dict], stage: str):
+    supabase = ctx.supabase
+    if tournament.get("status") == "COMPLETE":
+        st.error("Tournament is complete. Scores are locked.")
+        return
+    updated_any = False
+    for game_id, game in game_map.items():
+        if game.get("finalized_at"):
+            continue
+        score_a = int(st.session_state.get(f"{_score_key(stage, 'a', game_id)}", 0))
+        score_b = int(st.session_state.get(f"{_score_key(stage, 'b', game_id)}", 0))
+        if score_a == 0 and score_b == 0:
+            if game.get("score_a") or game.get("score_b"):
+                supabase.table("tournament_games").update({"score_a": None, "score_b": None}).eq("id", game_id).execute()
+                updated_any = True
+            continue
+        supabase.table("tournament_games").update({"score_a": score_a, "score_b": score_b}).eq("id", game_id).execute()
+        updated_any = True
+        try:
+            finalize_payload = finalize_game({**game, "score_a": score_a, "score_b": score_b})
+        except ValueError:
+            continue
+        match_payload = build_tournament_match_payload(tournament, game, teams_by_id, score_a=score_a, score_b=score_b)
+        process_matches(
+            [match_payload],
+            supabase=supabase,
+            club_id=str(ctx.club_id),
+            name_to_id=ctx.name_to_id,
+            df_players_all=ctx.df_players_all,
+            df_leagues=ctx.df_leagues,
+            df_meta=ctx.df_meta,
+        )
+        supabase.table("tournament_games").update(finalize_payload).eq("id", game_id).execute()
+
+        if stage == "PLAYOFF":
+            playoff_query = supabase.table("tournament_games").select("*").eq("tournament_id", tournament["id"]).eq("stage", "PLAYOFF")
+            if tournament.get("active_draw_id"):
+                playoff_query = playoff_query.eq("draw_id", tournament.get("active_draw_id"))
+            else:
+                playoff_query = playoff_query.is_("draw_id", "null")
+            playoff_games = (playoff_query.execute().data or [])
+            for update in resolve_playoff_dependencies(playoff_games):
+                supabase.table("tournament_games").update(update).eq("id", update["id"]).execute()
+
+    if updated_any:
+        st.success("Scores saved.")
+        st.session_state["force_data_refresh"] = True
+        st.rerun()
+
+
+def _score_key(stage: str, side: str, game_id: str) -> str:
+    prefix = "playoff" if stage == "PLAYOFF" else "rr"
+    return f"{prefix}_{side}_{game_id}"
