+++ streamlit_app.py
@@ -337,6 +337,8 @@ def main():
             tournament_manager,
             tournament_partner_board,
             tournament_registration,
+            tournament_live,
+            tournament_ops,
             tournaments,
             weekly_recap,
             weekly_recap_admin,
@@ -366,7 +368,9 @@ def main():
             "🔴 JUPR Live Admin": jupr_live_admin,
             "🎨 Theme QA": theme_gallery,
             "🏆 Tournaments": tournaments,
-            "🏆 Tournament Manager": tournament_manager,
+            "🛠️ Tournament Setup": tournament_manager,
+            "📋 Tournament Operations": tournament_ops,
+            "🔴 Tournament Live": tournament_live,
             "📝 Tournament Registration": tournament_registration,
             "🤝 Partner Board": tournament_partner_board,
             "🗞️ Weekly Recap": weekly_recap,
